requirements.txt
telethon==1.43.0
pyrogram==2.0.106
pymongo==4.8.0
motor==3.5.0
fastapi==0.115.0
uvicorn==0.30.0
cryptography>=43.0.0
aiofiles==24.1.0
pytest==8.3.0
pytest-asyncio==0.24.0
python-dotenv==1.0.1
redis==5.2.0
rarfile==4.2

# Machine Learning
scikit-learn==1.5.0
numpy==1.26.4
joblib==1.4.2
pandas==2.2.0

# Security
pyotp==2.9.0
fastapi-csrf-protect==0.3.0

# Cloud Storage
boto3==1.35.0

# Payment Gateways
requests==2.32.0
paytm-sdk==1.0.0

# Additional utilities
qrcode==8.0
pillow==10.4.0
setuptools>=75.0
aiohttp==3.10.0
markupsafe==2.1.5
